refactor: remove legacy defaults migration
