+++ TypeSwitchTests/AppFeatureTests.swift
@@ -3530,243 +3530,6 @@ final class AppFeatureTests: XCTestCase {
         XCTAssertEqual(state.totalSuccessfulSwitchCount, 12)
     }
 
-    func testLegacyRulesLoadedSavesBeforeMarkingMigrationCompleted() async {
-        let currentDate = Date(timeIntervalSince1970: 100)
-        let migrationDate = Date(timeIntervalSince1970: 777)
-        let migrationTracker = MigrationTracker()
-        let currentRule = AppRuleRecord(
-            bundleId: "com.test.notes",
-            lastKnownPath: "/Applications/Notes.app",
-            lastKnownName: "Notes",
-            strategy: .none,
-            createdAt: currentDate,
-            updatedAt: currentDate
-        )
-        let legacyRule = AppRuleRecord(
-            bundleId: currentRule.bundleId,
-            lastKnownPath: nil,
-            lastKnownName: currentRule.bundleId,
-            strategy: .fixed(inputMethodId: "ime.zh"),
-            createdAt: migrationDate,
-            updatedAt: migrationDate
-        )
-        let expectedRule = AppRuleRecord(
-            bundleId: currentRule.bundleId,
-            lastKnownPath: currentRule.lastKnownPath,
-            lastKnownName: currentRule.lastKnownName,
-            strategy: legacyRule.strategy,
-            createdAt: currentDate,
-            updatedAt: migrationDate
-        )
-        var initialState = makeMigrationTestState()
-        initialState.$appRulesStore.withLock {
-            $0.rules[currentRule.bundleId] = currentRule
-        }
-
-        let store = TestStore(initialState: initialState) {
-            AppFeature()
-        }
-        store.dependencies.appRulesStoreMigrationClient.save = { _ in
-            await migrationTracker.record(.save)
-        }
-        store.dependencies.legacyDefaultsMigrationClient.markCompleted = { version in
-            await migrationTracker.record(.markCompleted(version))
-        }
-
-        await store.send(.response(.legacyRulesLoaded(
-            [legacyRule.bundleId: legacyRule],
-            didCompleteLegacyMigration: false
-        ))) {
-            $0.$appRulesStore.withLock {
-                $0.rules[legacyRule.bundleId] = expectedRule
-            }
-        }
-        await store.finish()
-
-        let events = await migrationTracker.events
-        XCTAssertEqual(events, [.save, .markCompleted(2)])
-    }
-
-    func testLegacyRulesLoadedDoesNotMarkMigrationCompletedWhenSaveFails() async {
-        let migrationTracker = MigrationTracker()
-        let legacyRule = AppRuleRecord(
-            bundleId: "com.test.notes",
-            lastKnownPath: "/Applications/Notes.app",
-            lastKnownName: "Notes",
-            strategy: .fixed(inputMethodId: "ime.zh"),
-            createdAt: Date(timeIntervalSince1970: 777),
-            updatedAt: Date(timeIntervalSince1970: 777)
-        )
-
-        let store = TestStore(initialState: makeMigrationTestState()) {
-            AppFeature()
-        }
-        store.dependencies.appRulesStoreMigrationClient.save = { _ in
-            await migrationTracker.record(.save)
-            throw TestError.failed
-        }
-        store.dependencies.legacyDefaultsMigrationClient.markCompleted = { version in
-            await migrationTracker.record(.markCompleted(version))
-        }
-
-        await store.send(.response(.legacyRulesLoaded(
-            [legacyRule.bundleId: legacyRule],
-            didCompleteLegacyMigration: false
-        ))) {
-            $0.$appRulesStore.withLock {
-                $0.rules[legacyRule.bundleId] = legacyRule
-            }
-        }
-        await store.finish()
-
-        let events = await migrationTracker.events
-        XCTAssertEqual(events, [.save])
-    }
-
-    func testLegacyRulesLoadedWithoutLegacyMappingsKeepsCurrentRulesAndMarksCompleted() async {
-        let migrationTracker = MigrationTracker()
-        let currentRule = AppRuleRecord(
-            bundleId: "com.test.notes",
-            lastKnownPath: "/Applications/Notes.app",
-            lastKnownName: "Notes",
-            strategy: .ignored,
-            createdAt: Date(timeIntervalSince1970: 100),
-            updatedAt: Date(timeIntervalSince1970: 200)
-        )
-        var initialState = makeMigrationTestState()
-        initialState.$appRulesStore.withLock {
-            $0.rules[currentRule.bundleId] = currentRule
-        }
-
-        let store = TestStore(initialState: initialState) {
-            AppFeature()
-        }
-        store.dependencies.appRulesStoreMigrationClient.save = { _ in
-            await migrationTracker.record(.save)
-        }
-        store.dependencies.legacyDefaultsMigrationClient.markCompleted = { version in
-            await migrationTracker.record(.markCompleted(version))
-        }
-
-        await store.send(.response(.legacyRulesLoaded(
-            [:],
-            didCompleteLegacyMigration: true
-        )))
-        await store.finish()
-
-        let events = await migrationTracker.events
-        XCTAssertEqual(events, [.markCompleted(2)])
-        XCTAssertEqual(store.state.appRules[currentRule.bundleId], currentRule)
-    }
-
-    func testTaskMigratesWithExistingEmptyStoreBeforeScanningRunningAppsAndDoesNotRepeatV2Migration() async throws {
-        let migrationDate = Date(timeIntervalSince1970: 777)
-        let migrationTracker = MigrationTracker()
-        let rootDirectory = FileManager.default.temporaryDirectory
-            .appending(path: UUID().uuidString, directoryHint: .isDirectory)
-        let storeURL = rootDirectory.appending(path: AppStorageConfiguration.appRulesFilename)
-        try FileManager.default.createDirectory(
-            at: rootDirectory,
-            withIntermediateDirectories: true
-        )
-        try JSONEncoder().encode(AppRulesStore()).write(to: storeURL)
-        defer { try? FileManager.default.removeItem(at: rootDirectory) }
-
-        let legacyRule = AppRuleRecord(
-            bundleId: "com.test.notes",
-            lastKnownPath: "/Applications/Notes.app",
-            lastKnownName: "Notes",
-            strategy: .fixed(inputMethodId: "ime.zh"),
-            createdAt: migrationDate,
-            updatedAt: migrationDate
-        )
-
-        let initialState = AppFeature.State(
-            appRulesStore: Shared(
-                wrappedValue: AppRulesStore(),
-                .fileStorage(storeURL)
-            ),
-            appSwitchStatisticsStore: Shared(value: AppSwitchStatisticsStore()),
-            fallbackRuleStore: Shared(value: FallbackRuleStore())
-        )
-        XCTAssertTrue(FileManager.default.fileExists(atPath: storeURL.path))
-
-        let store = TestStore(initialState: initialState) {
-            AppFeature()
-        }
-        store.dependencies.date = .constant(migrationDate)
-        store.dependencies.legacyDefaultsMigrationClient.completedVersion = {
-            await migrationTracker.completedVersion
-        }
-        store.dependencies.legacyDefaultsMigrationClient.didCompleteLegacyMigration = { true }
-        store.dependencies.legacyDefaultsMigrationClient.loadRules = { receivedDate in
-            await migrationTracker.record(.loadRules(receivedDate))
-            return [legacyRule.bundleId: legacyRule]
-        }
-        store.dependencies.legacyDefaultsMigrationClient.markCompleted = { version in
-            await migrationTracker.markCompleted(version)
-        }
-        store.dependencies.appRulesStoreMigrationClient.save = { _ in
-            await migrationTracker.record(.save)
-        }
-        store.dependencies.workspaceClient.runningApplications = {
-            await migrationTracker.record(.runningApplications)
-            return []
-        }
-        store.dependencies.workspaceClient.events = finishedStream
-        store.dependencies.inputMethodClient.availabilityChanges = finishedStream
-        store.dependencies.inputMethodClient.selectionChanges = finishedStream
-
-        await store.send(.task) {
-            $0.nextInputMethodRefreshID = 1
-            $0.pendingInputMethodRefreshID = 0
-        }
-        await store.receive(.response(.legacyRulesLoaded(
-            [legacyRule.bundleId: legacyRule],
-            didCompleteLegacyMigration: true
-        ))) {
-            $0.$appRulesStore.withLock {
-                $0.rules[legacyRule.bundleId] = legacyRule
-            }
-        }
-        await receiveStartupResponses(from: store)
-        await store.finish()
-
-        let migratedRule = store.state.appRules[legacyRule.bundleId]
-        let firstRunEvents = await migrationTracker.events
-        XCTAssertLessThan(
-            try XCTUnwrap(firstRunEvents.firstIndex(of: .loadRules(migrationDate))),
-            try XCTUnwrap(firstRunEvents.firstIndex(of: .runningApplications))
-        )
-        XCTAssertEqual(firstRunEvents.filter { $0 == .save }.count, 1)
-        XCTAssertEqual(firstRunEvents.filter { $0 == .markCompleted(2) }.count, 1)
-
-        await store.send(.task) {
-            $0.inputMethodCatalogStatus = .loading
-            $0.nextInputMethodRefreshID = 2
-            $0.pendingInputMethodRefreshID = 1
-        }
-        await receiveStartupResponses(from: store)
-        await store.finish()
-
-        let secondRunEvents = await migrationTracker.events
-        XCTAssertEqual(secondRunEvents.filter {
-            if case .loadRules = $0 { return true }
-            return false
-        }.count, 1)
-        XCTAssertEqual(secondRunEvents.filter { $0 == .save }.count, 1)
-        XCTAssertEqual(secondRunEvents.filter { $0 == .markCompleted(2) }.count, 1)
-        XCTAssertEqual(store.state.appRules[legacyRule.bundleId], migratedRule)
-    }
-
-    private func makeMigrationTestState() -> AppFeature.State {
-        AppFeature.State(
-            appRulesStore: Shared(value: AppRulesStore()),
-            appSwitchStatisticsStore: Shared(value: AppSwitchStatisticsStore()),
-            fallbackRuleStore: Shared(value: FallbackRuleStore())
-        )
-    }
-
     private func makeRule(
         app: AppInfo,
         strategy: InputMethodStrategy,
@@ -3950,25 +3713,4 @@ private actor FrontmostApplicationGate {
 
 private enum TestError: Error {
     case failed
-}
-
-private actor MigrationTracker {
-    enum Event: Equatable {
-        case loadRules(Date)
-        case markCompleted(Int)
-        case runningApplications
-        case save
-    }
-
-    private(set) var completedVersion = 0
-    private(set) var events: [Event] = []
-
-    func record(_ event: Event) {
-        events.append(event)
-    }
-
-    func markCompleted(_ version: Int) {
-        completedVersion = version
-        events.append(.markCompleted(version))
-    }
 }
