--- TypeSwitchTests/AppRulesStoreMigrationTests.swift
@@ -1,405 +1,0 @@
-import Foundation
-@testable import TypeSwitch
-import XCTest
-
-final class AppRulesStoreMigrationTests: XCTestCase {
-    func testAppRulesStoreEncodingIncludesVersion() throws {
-        let store = AppRulesStore(
-            v: MigrationVersion.current,
-            rules: [
-                "com.test.notes": AppRuleRecord(
-                    bundleId: "com.test.notes",
-                    lastKnownPath: "/Applications/Notes.app",
-                    lastKnownName: "Notes",
-                    strategy: .fixed(inputMethodId: "ime.zh"),
-                    createdAt: Date(timeIntervalSince1970: 10),
-                    updatedAt: Date(timeIntervalSince1970: 20)
-                ),
-            ]
-        )
-
-        let data = try JSONEncoder().encode(store)
-        let payload = try XCTUnwrap(
-            JSONSerialization.jsonObject(with: data) as? [String: Any]
-        )
-
-        XCTAssertEqual(payload["v"] as? Int, MigrationVersion.current)
-        XCTAssertEqual(try JSONDecoder().decode(AppRulesStore.self, from: data), store)
-    }
-
-    func testAppRulesStoreRoundTripsIgnoredStrategy() throws {
-        let store = AppRulesStore(
-            rules: [
-                "com.test.passwords": AppRuleRecord(
-                    bundleId: "com.test.passwords",
-                    lastKnownPath: "/Applications/Passwords.app",
-                    lastKnownName: "Passwords",
-                    strategy: .ignored,
-                    strategyBeforeIgnoring: .followLast(lastInputMethodId: "ime.zh"),
-                    createdAt: Date(timeIntervalSince1970: 10),
-                    updatedAt: Date(timeIntervalSince1970: 20)
-                ),
-            ]
-        )
-
-        let data = try JSONEncoder().encode(store)
-
-        XCTAssertEqual(try JSONDecoder().decode(AppRulesStore.self, from: data), store)
-    }
-
-    func testLegacyIgnoredRuleWithoutPreviousStrategyDecodesWithNil() throws {
-        let data = Data(
-            #"{"bundleId":"com.test.passwords","lastKnownName":"Passwords","strategy":{"ignored":{}},"createdAt":0,"updatedAt":0}"#.utf8
-        )
-
-        let rule = try JSONDecoder().decode(AppRuleRecord.self, from: data)
-
-        XCTAssertEqual(rule.strategy, .ignored)
-        XCTAssertNil(rule.strategyBeforeIgnoring)
-    }
-
-    func testExistingInputMethodStrategiesStillDecode() throws {
-        let decoder = JSONDecoder()
-
-        XCTAssertEqual(
-            try decoder.decode(InputMethodStrategy.self, from: Data(#"{"none":{}}"#.utf8)),
-            .none
-        )
-        XCTAssertEqual(
-            try decoder.decode(InputMethodStrategy.self, from: Data(#"{"fixed":{"inputMethodId":"ime.zh"}}"#.utf8)),
-            .fixed(inputMethodId: "ime.zh")
-        )
-        XCTAssertEqual(
-            try decoder.decode(InputMethodStrategy.self, from: Data(#"{"followLast":{"lastInputMethodId":"ime.jp"}}"#.utf8)),
-            .followLast(lastInputMethodId: "ime.jp")
-        )
-    }
-
-    func testFallbackRuleStoreDefaultsAndEncodingIncludeVersion() throws {
-        let defaultStore = FallbackRuleStore()
-        XCTAssertEqual(defaultStore.v, MigrationVersion.current)
-        XCTAssertEqual(defaultStore.strategy, .none)
-
-        let store = FallbackRuleStore(strategy: .fixed(inputMethodId: "ime.abc"))
-        let data = try JSONEncoder().encode(store)
-        let payload = try XCTUnwrap(
-            JSONSerialization.jsonObject(with: data) as? [String: Any]
-        )
-
-        XCTAssertEqual(payload["v"] as? Int, MigrationVersion.current)
-        XCTAssertEqual(try JSONDecoder().decode(FallbackRuleStore.self, from: data), store)
-    }
-
-    func testAppSwitchStatisticsStoreDefaultsAndEncoding() throws {
-        let defaultStore = AppSwitchStatisticsStore()
-        XCTAssertTrue(defaultStore.counts.isEmpty)
-
-        let store = AppSwitchStatisticsStore(counts: [
-            "com.test.notes": 2,
-            "com.test.browser": 5,
-        ])
-        let data = try JSONEncoder().encode(store)
-
-        XCTAssertEqual(try JSONDecoder().decode(AppSwitchStatisticsStore.self, from: data), store)
-    }
-
-    func testAppRulesStoreURLUsesBundleIdentifierDirectory() {
-        let applicationSupportDirectory = URL(fileURLWithPath: "/tmp/Application Support", isDirectory: true)
-        let url = URL.appRulesStoreFileURL(
-            applicationSupportDirectory: applicationSupportDirectory,
-            bundleIdentifier: "com.test.TypeSwitch"
-        )
-
-        XCTAssertEqual(
-            url.path,
-            applicationSupportDirectory
-                .appending(path: "com.test.TypeSwitch", directoryHint: .isDirectory)
-                .appending(path: AppStorageConfiguration.appRulesFilename)
-                .path
-        )
-    }
-
-    func testFallbackRuleStoreURLUsesFallbackRuleFilename() {
-        let applicationSupportDirectory = URL(fileURLWithPath: "/tmp/Application Support", isDirectory: true)
-        let url = URL.fallbackRuleStoreFileURL(
-            applicationSupportDirectory: applicationSupportDirectory,
-            bundleIdentifier: "com.test.TypeSwitch"
-        )
-
-        XCTAssertEqual(AppStorageConfiguration.fallbackRuleFilename, "fallback-rule.json")
-        XCTAssertEqual(
-            url.path,
-            applicationSupportDirectory
-                .appending(path: "com.test.TypeSwitch", directoryHint: .isDirectory)
-                .appending(path: AppStorageConfiguration.fallbackRuleFilename)
-                .path
-        )
-        XCTAssertNotEqual(url, URL.appRulesStoreFileURL(
-            applicationSupportDirectory: applicationSupportDirectory,
-            bundleIdentifier: "com.test.TypeSwitch"
-        ))
-    }
-
-    func testAppSwitchStatisticsStoreURLUsesStatisticsFilename() {
-        let applicationSupportDirectory = URL(fileURLWithPath: "/tmp/Application Support", isDirectory: true)
-        let url = URL.appSwitchStatisticsStoreFileURL(
-            applicationSupportDirectory: applicationSupportDirectory,
-            bundleIdentifier: "com.test.TypeSwitch"
-        )
-
-        XCTAssertEqual(AppStorageConfiguration.appSwitchStatisticsFilename, "app-switch-statistics.json")
-        XCTAssertEqual(
-            url.path,
-            applicationSupportDirectory
-                .appending(path: "com.test.TypeSwitch", directoryHint: .isDirectory)
-                .appending(path: AppStorageConfiguration.appSwitchStatisticsFilename)
-                .path
-        )
-        XCTAssertNotEqual(url, URL.appRulesStoreFileURL(
-            applicationSupportDirectory: applicationSupportDirectory,
-            bundleIdentifier: "com.test.TypeSwitch"
-        ))
-        XCTAssertNotEqual(url, URL.fallbackRuleStoreFileURL(
-            applicationSupportDirectory: applicationSupportDirectory,
-            bundleIdentifier: "com.test.TypeSwitch"
-        ))
-    }
-
-    func testMergeKeepsCurrentRulesWhenLegacyMigrationPreviouslyCompleted() {
-        let currentDate = Date(timeIntervalSince1970: 100)
-        let legacyDate = Date(timeIntervalSince1970: 200)
-        let currentRules = [
-            "none": makeRule(bundleId: "none", strategy: .none, date: currentDate),
-            "fixed": makeRule(
-                bundleId: "fixed",
-                strategy: .fixed(inputMethodId: "current.fixed"),
-                date: currentDate
-            ),
-        ]
-        let legacyRules = [
-            "missing": makeRule(bundleId: "missing", strategy: .fixed(inputMethodId: "legacy.missing"), date: legacyDate),
-            "none": makeRule(bundleId: "none", strategy: .fixed(inputMethodId: "legacy.none"), date: legacyDate),
-        ]
-
-        let mergedRules = AppRulesStoreMigration.merge(
-            currentRules: currentRules,
-            legacyRules: legacyRules,
-            didCompleteLegacyMigration: true
-        )
-
-        XCTAssertEqual(mergedRules, currentRules)
-        XCTAssertNil(mergedRules["missing"])
-    }
-
-    func testMergeRecoversMissingAndUntouchedPlaceholderWhenLegacyMigrationWasIncomplete() {
-        let currentDate = Date(timeIntervalSince1970: 100)
-        let changedDate = Date(timeIntervalSince1970: 150)
-        let legacyDate = Date(timeIntervalSince1970: 200)
-        let placeholder = makeRule(bundleId: "placeholder", strategy: .none, date: currentDate)
-        var explicitDefault = makeRule(bundleId: "explicit", strategy: .none, date: currentDate)
-        explicitDefault.updatedAt = changedDate
-        let currentRules = [
-            placeholder.bundleId: placeholder,
-            explicitDefault.bundleId: explicitDefault,
-        ]
-        let legacyRules = [
-            "missing": makeRule(bundleId: "missing", strategy: .fixed(inputMethodId: "legacy.missing"), date: legacyDate),
-            "placeholder": makeRule(bundleId: "placeholder", strategy: .fixed(inputMethodId: "legacy.placeholder"), date: legacyDate),
-            "explicit": makeRule(bundleId: "explicit", strategy: .fixed(inputMethodId: "legacy.explicit"), date: legacyDate),
-        ]
-
-        let mergedRules = AppRulesStoreMigration.merge(
-            currentRules: currentRules,
-            legacyRules: legacyRules,
-            didCompleteLegacyMigration: false
-        )
-
-        XCTAssertEqual(mergedRules["missing"], legacyRules["missing"])
-        XCTAssertEqual(mergedRules["placeholder"]?.strategy, .fixed(inputMethodId: "legacy.placeholder"))
-        XCTAssertEqual(mergedRules["placeholder"]?.createdAt, currentDate)
-        XCTAssertEqual(mergedRules["placeholder"]?.updatedAt, legacyDate)
-        XCTAssertEqual(mergedRules["explicit"], explicitDefault)
-    }
-
-    func testMergeRecoversEmptyStoreEvenWhenLegacyMigrationPreviouslyCompleted() {
-        let legacyRule = makeRule(
-            bundleId: "missing",
-            strategy: .fixed(inputMethodId: "legacy.missing"),
-            date: Date(timeIntervalSince1970: 200)
-        )
-
-        let mergedRules = AppRulesStoreMigration.merge(
-            currentRules: [:],
-            legacyRules: [legacyRule.bundleId: legacyRule],
-            didCompleteLegacyMigration: true
-        )
-
-        XCTAssertEqual(mergedRules, [legacyRule.bundleId: legacyRule])
-    }
-
-    func testMergePreservesCurrentMetadataWhenRecoveringPlaceholder() {
-        let currentDate = Date(timeIntervalSince1970: 100)
-        let legacyDate = Date(timeIntervalSince1970: 200)
-        let currentRule = AppRuleRecord(
-            bundleId: "com.test.external",
-            lastKnownPath: "/Volumes/External/External.app",
-            lastKnownName: "External",
-            strategy: .none,
-            createdAt: currentDate,
-            updatedAt: currentDate
-        )
-        let legacyRule = AppRuleRecord(
-            bundleId: currentRule.bundleId,
-            lastKnownPath: nil,
-            lastKnownName: currentRule.bundleId,
-            strategy: .fixed(inputMethodId: "legacy.external"),
-            createdAt: legacyDate,
-            updatedAt: legacyDate
-        )
-
-        let mergedRules = AppRulesStoreMigration.merge(
-            currentRules: [currentRule.bundleId: currentRule],
-            legacyRules: [legacyRule.bundleId: legacyRule],
-            didCompleteLegacyMigration: false
-        )
-
-        XCTAssertEqual(
-            mergedRules[currentRule.bundleId],
-            AppRuleRecord(
-                bundleId: currentRule.bundleId,
-                lastKnownPath: currentRule.lastKnownPath,
-                lastKnownName: currentRule.lastKnownName,
-                strategy: legacyRule.strategy,
-                createdAt: currentDate,
-                updatedAt: legacyDate
-            )
-        )
-    }
-
-    func testMergeRecoversMissingMetadataFromMatchedLegacyRule() {
-        let currentDate = Date(timeIntervalSince1970: 100)
-        let legacyDate = Date(timeIntervalSince1970: 200)
-        let currentRule = AppRuleRecord(
-            bundleId: "com.test.editor",
-            lastKnownPath: nil,
-            lastKnownName: "com.test.editor",
-            strategy: .none,
-            createdAt: currentDate,
-            updatedAt: currentDate
-        )
-        let legacyRule = AppRuleRecord(
-            bundleId: currentRule.bundleId,
-            lastKnownPath: "/Applications/Editor.app",
-            lastKnownName: "Editor",
-            strategy: .fixed(inputMethodId: "legacy.fixed"),
-            createdAt: legacyDate,
-            updatedAt: legacyDate
-        )
-
-        let mergedRules = AppRulesStoreMigration.merge(
-            currentRules: [currentRule.bundleId: currentRule],
-            legacyRules: [legacyRule.bundleId: legacyRule],
-            didCompleteLegacyMigration: false
-        )
-
-        XCTAssertEqual(
-            mergedRules[currentRule.bundleId],
-            AppRuleRecord(
-                bundleId: currentRule.bundleId,
-                lastKnownPath: legacyRule.lastKnownPath,
-                lastKnownName: legacyRule.lastKnownName,
-                strategy: legacyRule.strategy,
-                createdAt: currentDate,
-                updatedAt: legacyDate
-            )
-        )
-    }
-
-    func testMergeRestoresNoneRuleWithoutDiscardingCurrentMetadataWhenLegacyAppIsUnmatched() throws {
-        let currentDate = Date(timeIntervalSince1970: 100)
-        let legacyDate = Date(timeIntervalSince1970: 200)
-        let appURL = FileManager.default.temporaryDirectory
-            .appending(path: "Editor-\(UUID().uuidString).app", directoryHint: .isDirectory)
-        try FileManager.default.createDirectory(at: appURL, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: appURL) }
-
-        let currentRule = AppRuleRecord(
-            bundleId: "com.test.editor",
-            lastKnownPath: appURL.path,
-            lastKnownName: "Editor",
-            strategy: .none,
-            createdAt: currentDate,
-            updatedAt: currentDate
-        )
-        let legacyRule = AppRuleRecord(
-            bundleId: currentRule.bundleId,
-            lastKnownPath: nil,
-            lastKnownName: currentRule.bundleId,
-            strategy: .fixed(inputMethodId: "legacy.fixed"),
-            createdAt: legacyDate,
-            updatedAt: legacyDate
-        )
-
-        let mergedRules = AppRulesStoreMigration.merge(
-            currentRules: [currentRule.bundleId: currentRule],
-            legacyRules: [legacyRule.bundleId: legacyRule],
-            didCompleteLegacyMigration: false
-        )
-        let mergedRule = try XCTUnwrap(mergedRules[currentRule.bundleId])
-
-        XCTAssertEqual(
-            mergedRule,
-            AppRuleRecord(
-                bundleId: currentRule.bundleId,
-                lastKnownPath: currentRule.lastKnownPath,
-                lastKnownName: currentRule.lastKnownName,
-                strategy: legacyRule.strategy,
-                createdAt: currentDate,
-                updatedAt: legacyDate
-            )
-        )
-        XCTAssertTrue(mergedRule.isAvailable)
-    }
-
-    func testMergeIsIdempotent() {
-        let currentDate = Date(timeIntervalSince1970: 100)
-        let legacyDate = Date(timeIntervalSince1970: 200)
-        let currentRules = [
-            "none": makeRule(bundleId: "none", strategy: .none, date: currentDate),
-            "fixed": makeRule(bundleId: "fixed", strategy: .fixed(inputMethodId: "current.fixed"), date: currentDate),
-        ]
-        let legacyRules = [
-            "none": makeRule(bundleId: "none", strategy: .fixed(inputMethodId: "legacy.none"), date: legacyDate),
-            "fixed": makeRule(bundleId: "fixed", strategy: .fixed(inputMethodId: "legacy.fixed"), date: legacyDate),
-        ]
-
-        let firstMerge = AppRulesStoreMigration.merge(
-            currentRules: currentRules,
-            legacyRules: legacyRules,
-            didCompleteLegacyMigration: false
-        )
-        let secondMerge = AppRulesStoreMigration.merge(
-            currentRules: firstMerge,
-            legacyRules: legacyRules,
-            didCompleteLegacyMigration: false
-        )
-
-        XCTAssertEqual(secondMerge, firstMerge)
-        XCTAssertEqual(secondMerge["fixed"]?.updatedAt, currentDate)
-    }
-
-    private func makeRule(
-        bundleId: String,
-        strategy: InputMethodStrategy,
-        date: Date
-    ) -> AppRuleRecord {
-        AppRuleRecord(
-            bundleId: bundleId,
-            lastKnownPath: "/Applications/\(bundleId).app",
-            lastKnownName: bundleId,
-            strategy: strategy,
-            createdAt: date,
-            updatedAt: date
-        )
-    }
-}
